# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /artists:
    get:
      operationId: listArtists
      tags: [artists]
      summary: List all artists
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: leadTier
          in: query
          schema:
            type: string
        - name: outreachStatus
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: followersEstimate
          in: query
          schema:
            type: string
        - name: engagementLevel
          in: query
          schema:
            type: string
        - name: labelStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
    post:
      operationId: createArtist
      tags: [artists]
      summary: Create artist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
  /artists/{id}:
    get:
      operationId: getArtist
      tags: [artists]
      summary: Get artist by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
    put:
      operationId: updateArtist
      tags: [artists]
      summary: Update artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
    delete:
      operationId: deleteArtist
      tags: [artists]
      summary: Delete artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /artist-saved-views:
    get:
      operationId: listArtistSavedViews
      tags: [artists]
      summary: List saved views for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Saved view list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtistSavedView"
    post:
      operationId: createArtistSavedView
      tags: [artists]
      summary: Create a saved view
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistSavedViewBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistSavedView"
  /artist-saved-views/{id}:
    delete:
      operationId: deleteArtistSavedView
      tags: [artists]
      summary: Delete a saved view
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ARTIST DUPLICATES ──────────────────────────────────────────────────
...
  /artists/graph:
    get:
      operationId: getArtistGraph
      tags: [artists]
      summary: Get artist network graph data
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Graph data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistGraphData"
...
  /artists/territory-stats:
    get:
      operationId: getArtistTerritoryStats
      tags: [artists]
      summary: Get artist territory statistics grouped by city
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Territory stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistTerritoryStats"
  /artists/{id}/relationships:
    get:
      operationId: listArtistRelationships
      tags: [artists]
      summary: List relationships for an artist
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Relationship list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtistRelationship"
    post:
      operationId: createArtistRelationship
      tags: [artists]
      summary: Create an artist relationship
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistRelationshipBody"
      responses:
        "201":
          description: Relationship created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistRelationship"
  /artists/{id}/relationships/{relId}:
    delete:
      operationId: deleteArtistRelationship
      tags: [artists]
      summary: Delete an artist relationship
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: relId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── DISCOVERY ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ArtistBody:
      type: object
      properties:
        name:
          type: string
        genre:
          type: string
          nullable: true
        labelStatus:
          type: array
          items:
            type: string
        bio:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        outreachStatus:
          type: string
          enum: [new, contacted, in_talks, signed, passed]
        revenuePotential:
          type: string
          nullable: true
        followersEstimate:
          type: string
          nullable: true
        engagementLevel:
          type: string
          enum: [low, medium, high]
          nullable: true
        streamingLinks:
          type: object
          additionalProperties:
            type: string
        socialLinks:
          type: object
          additionalProperties:
            type: string
        tags:
          type: array
          items:
            type: string
        contactId:
          type: integer
          nullable: true
        spotifyId:
          type: string
          nullable: true
        youtubeChannelId:
          type: string
          nullable: true
        originCity:
          type: string
          nullable: true
        originState:
          type: string
          nullable: true
        originCountry:
          type: string
          nullable: true
        photoUrls:
          type: array
          items:
            type: string
      required: [name]
    Artist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        genre:
          type: string
          nullable: true
        labelStatus:
          type: array
          items:
            type: string
        bio:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        lat:
          type: number
          format: float
          nullable: true
        lng:
          type: number
          format: float
          nullable: true
        outreachStatus:
          type: string
        revenuePotential:
          type: string
          nullable: true
        followersEstimate:
          type: string
          nullable: true
        engagementLevel:
          type: string
          enum: [low, medium, high]
          nullable: true
        imageUrl:
          type: string
          nullable: true
        spotifyId:
          type: string
          nullable: true
        youtubeChannelId:
          type: string
          nullable: true
        originCity:
          type: string
          nullable: true
        originState:
          type: string
          nullable: true
        originCountry:
          type: string
          nullable: true
        photoUrls:
          type: array
          items:
            type: string
        streamingLinks:
          type: object
          additionalProperties:
            type: string
        socialLinks:
          type: object
          additionalProperties:
            type: string
        tags:
          type: array
          items:
            type: string
        contactId:
          type: integer
          nullable: true
        aiAnalysis:
          $ref: "#/components/schemas/ArtistAiAnalysis"
          nullable: true
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, labelStatus, outreachStatus, streamingLinks, socialLinks, tags, createdBy, createdAt, updatedAt]
    ArtistAiAnalysis:
      type: object
      properties:
        id:
          type: integer
        artistId:
          type: integer
        summary:
          type: string
        brandingScore:
          type: integer
        growthScore:
          type: integer
        professionalismScore:
          type: integer
        leadTier:
          type: string
        recommendations:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, artistId, summary, brandingScore, growthScore, professionalismScore, leadTier, recommendations, createdAt, updatedAt]
...
    ArtistSavedView:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        filters:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, filters, createdAt]
...
    ArtistSavedViewBody:
      type: object
      properties:
        name:
          type: string
        filters:
          type: object
          additionalProperties: true
      required: [name, filters]

    # ── SEARCH SCHEMAS ────────────────────────────────────────────────────
...
    ArtistRelationship:
      type: object
      properties:
        id:
          type: integer
        fromArtistId:
          type: integer
        toEntityId:
          type: integer
          nullable: true
        toEntityType:
          type: string
          enum: [artist, producer, engineer, venue, label]
        toEntityName:
          type: string
          nullable: true
        relationshipType:
          type: string
          enum: [collaborator, producer, engineer, venue, label, other]
        notes:
          type: string
          nullable: true
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, fromArtistId, toEntityType, relationshipType, createdBy, createdAt]
    ArtistRelationshipBody:
      type: object
      properties:
        toEntityId:
          type: integer
          nullable: true
        toEntityType:
          type: string
          enum: [artist, producer, engineer, venue, label]
        toEntityName:
          type: string
          nullable: true
        relationshipType:
          type: string
          enum: [collaborator, producer, engineer, venue, label, other]
        notes:
          type: string
          nullable: true
      required: [relationshipType]
    ArtistGraphData:
      type: object
      properties:
        nodes:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              genre:
                type: string
                nullable: true
              city:
                type: string
                nullable: true
              state:
                type: string
                nullable: true
              labelStatus:
                type: array
                items:
                  type: string
              outreachStatus:
                type: string
            required: [id, name, labelStatus, outreachStatus]
        links:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              source:
                type: integer
              target:
                type: integer
              type:
                type: string
            required: [id, source, target, type]
        allRelationships:
          type: array
          items:
            $ref: "#/components/schemas/ArtistRelationship"
      required: [nodes, links, allRelationships]
...
    ArtistTerritoryStats:
      type: object
      properties:
        cities:
          type: array
          items:
            type: object
            properties:
              city:
                type: string
              state:
                type: string
                nullable: true
              country:
                type: string
                nullable: true
              lat:
                type: number
                nullable: true
              lng:
                type: number
                nullable: true
              count:
                type: integer
              labelBreakdown:
                type: object
                additionalProperties:
                  type: integer
              outreachBreakdown:
                type: object
                additionalProperties:
                  type: integer
            required: [city, count, labelBreakdown, outreachBreakdown]
        total:
          type: integer
        withCity:
          type: integer
        geocoded:
          type: integer
      required: [cities, total, withCity, geocoded]